requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
rapidfuzz
python-dotenv
fastapi
pytest
ruff